Revert "feat: custom user agent InternalCognitoUserPool (#11716)"
This reverts commit 94bfc5f5bce7fab723591d99daf7cd9185443b89.
