+++ packages/amazon-cognito-identity-js/__tests__/internalsIndex.test.js
@@ -1,0 +1,13 @@
+import * as exported from '../src/internals/index';
+
+describe('import * keys', () => {
+	it('should match snapshot', () => {
+		expect(Object.keys(exported)).toMatchInlineSnapshot(`
+		Array [
+		  "addAuthCategoryToCognitoUserAgent",
+		  "addFrameworkToCognitoUserAgent",
+		  "InternalCognitoUser",
+		]
+	`);
+	});
+});
