+++ packages/amazon-cognito-identity-js/index.d.ts
@@ -1,4 +1,4 @@
-import { InternalCognitoUser, InternalCognitoUserPool } from './internals';
+import { InternalCognitoUser } from './internals';
 declare module 'amazon-cognito-identity-js' {
 	//import * as AWS from "aws-sdk";
 
@@ -320,7 +320,18 @@ declare module 'amazon-cognito-identity-js' {
 		AdvancedSecurityDataCollectionFlag?: boolean;
 	}
 
-	export class CognitoUserPool extends InternalCognitoUserPool {
+	export class CognitoUserPool {
+		constructor(
+			data: ICognitoUserPoolData,
+			wrapRefreshSessionCallback?: (
+				target: NodeCallback.Any
+			) => NodeCallback.Any
+		);
+
+		public getUserPoolId(): string;
+		public getUserPoolName(): string;
+		public getClientId(): string;
+
 		public signUp(
 			username: string,
 			password: string,
@@ -329,6 +340,8 @@ declare module 'amazon-cognito-identity-js' {
 			callback: NodeCallback<Error, ISignUpResult>,
 			clientMetadata?: ClientMetadata
 		): void;
+
+		public getCurrentUser(): CognitoUser | null;
 	}
 
 	export interface ICognitoUserSessionData {
