+++ packages/amazon-cognito-identity-js/src/internals/index.js
@@ -4,4 +4,3 @@ export {
 } from '../UserAgent';
 
 export { InternalCognitoUser } from './InternalCognitoUser';
-export { InternalCognitoUserPool } from './InternalCognitoUserPool';
